Split all API implementations to separate files.
Also organized the files hierarchically to be able to find them more easily.

Remote commands have been placed in their own directory, with RemoteCommandStatus in the root of that directory.
